+++ Gemfile
@@ -1,0 +1,6 @@
+source 'https://rubygems.org'
+
+gem 'memory_profiler'
+gem 'ruby-prof'
+gem 'stackprof'
+gem 'rspec'
